+++ R/getCDS.R
@@ -52,9 +52,7 @@ getCDS <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
         )
         
         if (nrow(FoundOrganism) == 0) {
-            cat("\n")
-            cat(paste0("----------> No reference genome or representative genome was found for '",organism,"'. Thus, download for this species has been omitted."))
-            cat("\n")
+            message(paste0("----------> No reference genome or representative genome was found for '",organism,"'. Thus, download for this species has been omitted."))
         } else { 
             if (nrow(FoundOrganism) > 1) {
                 warnings(
@@ -81,6 +79,11 @@ getCDS <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
             
             local.org <- stringr::str_replace_all(organism,"-","_")
             local.org <- stringr::str_replace_all(organism,"\\/","_")
+            
+            if (!exists.ftp.file(url = paste0(FoundOrganism$ftp_path,"/"), file.path = download_url)) {
+                message("Unfortunately no CDS file could be found for organism '",organism,"'. Thus, the download of this organism has been omitted.")
+                return(FALSE) 
+            }
             
             if (nrow(FoundOrganism) == 1) {
                 tryCatch({
@@ -177,13 +180,9 @@ getCDS <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
                 stop(
                     "The API 'http://rest.ensembl.org' does not seem to work properly. Are you connected to the internet? Is the homepage 'http://rest.ensembl.org' currently available?", call. = FALSE
                 ))
-            
-            cwd <- getwd()
-            
-            setwd(path)
-            
+        
             # generate CDS documentation
-            sink(paste0("doc_",new.organism,"_db_",db,".txt"))
+            sink(file.path(path, paste0("doc_",new.organism,"_db_",db,".txt")))
             
             cat(paste0("File Name: ", cds.path))
             cat("\n")
@@ -204,8 +203,6 @@ getCDS <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
             cat(paste0("genebuild_initial_release_date: ", json.qry.info$genebuild_initial_release_date))
             
             sink()
-            
-            setwd(cwd)
             
             print(
                 paste0(
@@ -254,12 +251,8 @@ getCDS <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
                     "The API 'http://rest.ensemblgenomes.org' does not seem to work properly. Are you connected to the internet? Is the homepage 'http://rest.ensemblgenomes.org' currently available?", call. = FALSE
                 ))
             
-            cwd <- getwd()
-            
-            setwd(path)
-            
             # generate CDS documentation
-            sink(paste0("doc_",new.organism,"_db_",db,".txt"))
+            sink(file.path(path, paste0("doc_",new.organism,"_db_",db,".txt")))
             
             cat(paste0("File Name: ", cds.path))
             cat("\n")
@@ -280,9 +273,7 @@ getCDS <- function(db = "refseq", organism, path = file.path("_ncbi_downloads","
             cat(paste0("genebuild_initial_release_date: ", json.qry.info$genebuild_initial_release_date))
             
             sink()
-            
-            setwd(cwd)
-            
+        
             print(
                 paste0(
                     "The CDS of '",