+++ R/getGenome.R
@@ -78,9 +78,7 @@ getGenome <-
                 )
             
             if (nrow(FoundOrganism) == 0) {
-                cat("\n")
-                cat(paste0("----------> No reference genome or representative genome was found for '",organism,"'. Thus, download for this species has been omitted."))
-                cat("\n")
+                message(paste0("----------> No reference genome or representative genome was found for '",organism,"'. Thus, download for this species has been omitted."))
             } else {
                 if (nrow(FoundOrganism) > 1) {
                     warnings(
@@ -104,6 +102,11 @@ getGenome <-
                             "_genomic.fna.gz"
                         )
                     )
+                
+                if (!exists.ftp.file(url = paste0(FoundOrganism$ftp_path,"/"), file.path = download_url)) {
+                   message("Unfortunately no genome file could be found for organism '",organism,"'. Thus, the download of this organism has been omitted.")
+                   return(FALSE) 
+                }
                 
                 # download_url <- paste0(query$ftp_path,query$`# assembly_accession`,"_",query$asm_name,"_genomic.fna.gz")
                 
@@ -196,12 +199,8 @@ getGenome <-
                         "The API 'http://rest.ensembl.org' does not seem to work properly. Are you connected to the internet? Is the homepage 'http://rest.ensembl.org' currently available?", call. = FALSE
                     ))
                 
-                cwd <- getwd()
-                
-                setwd(path)
-                
                 # generate Genome documentation
-                sink(paste0("doc_",new.organism,"_db_",db,".txt"))
+                sink(file.path(path, paste0("doc_",new.organism,"_db_",db,".txt")))
                 
                 cat(paste0("File Name: ", genome.path))
                 cat("\n")
@@ -222,8 +221,6 @@ getGenome <-
                 cat(paste0("genebuild_initial_release_date: ", json.qry.info$genebuild_initial_release_date))
                 
                 sink()
-                
-                setwd(cwd)
                 
                 print(
                     paste0(
@@ -270,13 +267,9 @@ getGenome <-
                     stop(
                         "The API 'http://rest.ensemblgenomes.org' does not seem to work properly. Are you connected to the internet? Is the homepage 'http://rest.ensemblgenomes.org' currently available?", call. = FALSE
                     ))
-                
-                cwd <- getwd()
-                
-                setwd(path)
-                
+            
                 # generate Genome documentation
-                sink(paste0("doc_",new.organism,"_db_",db,".txt"))
+                sink(file.path(path,paste0("doc_",new.organism,"_db_",db,".txt")))
                 
                 cat(paste0("File Name: ", genome.path))
                 cat("\n")
@@ -297,8 +290,6 @@ getGenome <-
                 cat(paste0("genebuild_initial_release_date: ", json.qry.info$genebuild_initial_release_date))
                 
                 sink()
-                
-                setwd(cwd)
                 
                 print(
                     paste0(